+++ src/Pho/Server/Rest/Routes/Authentication.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    ["GET", "/login", "login"],
+    ["GET", "/signup", "signup"],
+    ["GET", "/logout", "logout"]
+];